# Vectorize RMSNorm CUDA kernel