@bbeckca 2025-08-17
Update: Performed e2e benchmark for latency using x2 H100 and observed ~4.8% reduction with vectorized rms_norm_kernel without shared memory. I'm planning to benchmark further with other models/configurations to understand this better. Let me know if you folks have any thoughts/preferences.

```
VLLM_USE_V1=1 buck run //vllm:benchmark_latency --   --model facebook/opt-125m   --input-len 32 --output-len 128   --dtype float16 --batch-size 128 --enforce-eager

baseline
Avg latency: 0.6731001150328666 seconds
10% percentile latency: 0.6315004411269911 seconds
25% percentile latency: 0.6374467767454917 seconds
50% percentile latency: 0.6609808525245171 seconds
75% percentile latency: 0.6663343247637386 seconds
90% percentile latency: 0.6783564629906322 seconds
99% percentile latency: 1.0606678462994754 seconds

vectorized
Avg latency: 0.6741364902312247 seconds
10% percentile latency: 0.6488532381190453 seconds
25% percentile latency: 0.6496474424784537 seconds
50% percentile latency: 0.6550397415121552 seconds
75% percentile latency: 0.6627388229971984 seconds
90% percentile latency: 0.666680197842652 seconds
99% percentile latency: 1.039106485189987 seconds

vectorized (without shared memory)
Avg latency: 0.6409522683670124 seconds
10% percentile latency: 0.6146808989578858 seconds
25% percentile latency: 0.6191042929858668 seconds
50% percentile latency: 0.6249049424950499 seconds
75% percentile latency: 0.6411205744952895 seconds
90% percentile latency: 0.6505832085269503 seconds
99% percentile latency: 0.8902701860625533 seconds
```