@bbeckca 2025-08-17
Ran quick E2E latency on a few models (bs=128, in=32, out=128). Overall: small win on Llama-3.2, neutral on Qwen fp16, and Gemma-2-9B works better if we skip the shared-mem cached row. I can rerun with further settings or add other models if there's any preferences.

@yewentao256  Wondering if we should consider adding a knob to disable the cached-row path where it regresses for some models?