@bbeckca 2025-08-20
> Ran quick E2E latency on a few models (bs=128, in=32, out=128). Overall: small win on Llama-3.2, neutral on Qwen fp16, and Gemma-2-9B works better if we skip the shared-mem cached row. I can rerun with further settings or add other models if there's any preferences.
> 
> @yewentao256 Wondering if we should consider adding a knob to disable the cached-row path where it regresses for some models?
> 
> Model	dtype	baseline (s)	vectorized (s)	Δ vs base	no-shared-mem (s)	Δ vs base
> Llama-3.2-3B-Instruct	fp16	1.2895	1.2698	**−1.5%**	1.2863	−0.3%
> Llama-3.2-3B-Instruct	bf16	1.2772	1.2663	**−0.9%**	1.2892	+0.9%
> Qwen2-7B	fp16	1.7673	1.7692	+0.1%	1.7679	+0.0%
> Qwen2-7B	bf16	1.7307	1.7143	**−0.9%**	1.6929	**−2.2%**
> Gemma-2-9B	bf16	4.1187	4.2222	**+2.5%**	4.0485	**−1.7%**

cc @mgoin @WoosukKwon for any additional thoughts.