@bbeckca 2025-09-14
> Sorry for the late response, there are some problems with my github notifications but now fixed. The overall work I think is good, could you
> 
> 1. test the accuracy using `lm_eval`?
> 2. benchmark the kernel directly?
>    One example could be seen here: #22036

Thanks for your patience. I’ve updated the test plan following your guidance. There are some regressions at very small sizes, the kernel shows significant performance gains for 1k+ tokens. Could you take another look?