@bbeckca 2025-09-22
@yewentao256 Heads up, I’m hitting some local env issues so my RMSNorm revisions are taking longer than expected. Working through it and will follow up once it's ready.